atlas console without dropwizard dependency
* separate atlas server into dropwizard and non-dropwizard parts

* rely on dropwizard-jackson to get kvs service discovery

* centralize AtlasDbConfig loading
