+++ atlasdb-server-dropwizard/.gitignore
