+++ atlasdb-server-dropwizard/src/test/resources/logback.xml
