+++ settings.gradle
@@ -21,6 +21,7 @@ include ":atlasdb-spi"
 include ":atlasdb-commons"
 include ":atlasdb-partition-manager"
 include ":atlasdb-server"
+include ":atlasdb-server-dropwizard"
 include ":atlasdb-server-with-endpoint"
 include ":commons-cassandra"
 include ":commons-executors"
